aoc_2022/requirements.txt
numpy
pyperclip
pytest
python-dotenv
tqdm
requests
